Add files via upload
An exploratory data analysis of the  Unicorn companies dataset was examined using the univariate, bivariate and and multivariate analysis. The data set was imported as a csv file and processed using various libraries such as numpy and pandas then visualised using the matplotlib.pyplot and seaborn. Data cleaning was the initial step whereby missing values and wrong data types were identified and then corrected. The 'city' had 16 missing values while 'select investors' had only one. The rest of the data was complete. Valuation was also processed as a string and was converted into a float. However, i had a challenge in converting funding to a float hence, the analysis was carried out without this column as i could not run it. Missing values were visualised with a heatmap. Data checking was further performed on columns using the categorical and numerical methods. 
The numerical column 'Year Founded' was described generating a mean year of 2012, standard deviation of 5 years for companies to belong to the unicorn, the minimum year of joining being 1919 and the maximum 2021.  Industry Valuation was also described generating results with the mean of $1B and Fintech being the highest performer. 
The univariate, bivariate and mutivariate analysis was done and the graphs generated included in the notebook.